onEscKeydown перенесен в util.js
