+++ js/big-picture.js
@@ -1,4 +1,4 @@
-import { isEscapeKey } from './util.js';
+import { onEscKeydown } from './util.js';
 import { indefications, body } from './const.js';
 import { renderCommentsList } from './render-comments.js';
 
@@ -6,13 +6,6 @@ const bigPicture = document.querySelector('.big-picture');
 const commentsLoader = bigPicture.querySelector('.comments-loader');
 //const socialComments = bigPicture.querySelector('.social__comment-count');
 const closeButtonPicture = bigPicture.querySelector('.big-picture__cancel');
-
-const onEscKeydown = (evt) => {
-  if (isEscapeKey(evt)) {
-    evt.preventDefault();
-    closeBigPicture();
-  }
-};
 
 const renderPictureDetails = ({ url, likes, description, comments }) => {
   bigPicture.querySelector('.big-picture__img img').src = url;
