+++ js/util.js
@@ -1,3 +1,6 @@
+import { closeUploadOverlay } from './form.js';
+import { closeBigPicture } from './big-picture.js';
+
 const getRandomInteger = (a, b) => {
   const lower = Math.ceil(Math.min(a, b));
   const upper = Math.floor(Math.max(a, b));
@@ -23,4 +26,16 @@ const isEscapeKey = (evt) => evt.key === 'Escape';
 
 const isEnterKey = (evt) => evt.key === 'Enter';
 
-export { getRandomInteger, getRandomIdentifier, getRandomArrayElement, isEscapeKey, isEnterKey };
+const onEscKeydown = (evt) => {
+  const inputFocus = evt.target.matches('input.text__hashtags:focus') || evt.target.matches('textarea.text__description:focus');
+  if (inputFocus) {
+    return false; // - если фокус находится в поле ввода комментария или хештега, нажатие на Esc не срабатывает.
+  }
+  if (isEscapeKey(evt)) {
+    evt.preventDefault();
+    closeUploadOverlay();
+    closeBigPicture();
+  }
+};
+
+export { getRandomInteger, getRandomIdentifier, getRandomArrayElement, isEnterKey, onEscKeydown };
